Correct a crash on 64 bit systems.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@26528 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSLayoutManager.m
@@ -674,8 +674,8 @@ Fills in all glyph holes up to last. only looking at levels below level
             }
           else
             {
-              unsigned int cindex = cpos;
-              unsigned int gindex = gpos + total;
+              NSUInteger cindex = cpos;
+              NSUInteger gindex = gpos + total;
               
               // Cache the current run
               cached_run = (glyph_run_t *)h;
